requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alpaca-py==0.42.0
annotated-types==0.7.0
appnope==0.1.4
asttokens==3.0.0
asyncio==3.4.3
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
comm==0.2.2
contourpy==1.3.2
cryptography==45.0.5
curl_cffi==0.12.0
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
discord.py==2.5.2
discord_colorize==0.0.8
executing==2.2.0
fonttools==4.59.0
frozendict==2.4.6
frozenlist==1.7.0
idna==3.10
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
matplotlib==3.10.3
matplotlib-inline==0.1.7
msgpack==1.1.1
multidict==6.6.3
multitasking==0.0.12
nest-asyncio==1.6.0
numpy==2.3.1
packaging==25.0
pandas==2.3.1
parso==0.8.4
peewee==3.18.2
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyzmq==27.0.0
requests==2.32.4
schwabdev==2.5.0
six==1.17.0
soupsieve==2.7
sseclient-py==1.8.0
stack-data==0.6.3
tornado==6.5.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.1
yfinance==0.2.65
